Add the proxy resolver perftest to GYP.
BUG=2764

Review URL: http://codereview.chromium.org/39239

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@11082 0039d316-1c4b-4281-b951-d872f2087c98

+++ net/proxy/proxy_resolver_perftest.cc
@@ -23,7 +23,7 @@ struct PacQuery {
 };
 
 // Entry listing which PAC scripts to load, and which URLs to try resolving.
-// |queries| should be terminated by {NULL, NULL}. A sentinel is used 
+// |queries| should be terminated by {NULL, NULL}. A sentinel is used
 // rather than a length, to simplify using initializer lists.
 struct PacPerfTest {
   const char* pac_name;
@@ -57,7 +57,7 @@ static PacPerfTest kPerfTests[] = {
 };
 
 int PacPerfTest::NumQueries() const {
-  for (int i = 0; i < arraysize(queries); ++i) {
+  for (size_t i = 0; i < arraysize(queries); ++i) {
     if (queries[i].query_url == NULL)
       return i;
   }
